+++ frontend/components/screens/schedule/create/schedule-date/schedule-date.interface.ts
@@ -5,4 +5,7 @@ export interface ScheduleDateProps {
 	isScheduleName: string;
 	onChangeStartEndPeriod: (startPeriod: string, endPeriod: string) => void;
 	onChangePeriods: (dates: PeriodsType[]) => void;
+	handleChangeDate: (dates: [Date, Date]) => void;
+	startDate: Date;
+	endDate: Date;
 }
