+++ frontend/components/screens/schedule/create/schedule-period/schedule-period.interface.ts
@@ -3,11 +3,7 @@ import { Union, schdulePages } from 'types';
 
 export interface SchedulePeriodProps {
 	onChangePage: (page: Union<typeof schdulePages>) => void;
-	onChangePeriods: (dates: PeriodsType[]) => void;
-	onChangeScheduleName: (name: string) => void;
-	onChangeStartEndPeriod: (startPeriod: string, endPeriod: string) => void;
 	isPeriods: PeriodsType[];
-	isScheduleName: string;
-	updateStartDate?: Date;
-	updateEndDate?: Date;
+	startDate: Date;
+	endDate: Date;
 }
